+++ opentelemetry-exporter-gcp-logging/src/opentelemetry/exporter/cloud_logging/__init__.py
@@ -18,7 +18,7 @@ import datetime
 import json
 import logging
 import urllib.parse
-from typing import Any, Mapping, Optional, Sequence
+from typing import Any, Mapping, MutableMapping, Optional, Sequence
 
 import google.auth
 from google.api.monitored_resource_pb2 import (  # pylint: disable = no-name-in-module
@@ -50,6 +50,7 @@ from opentelemetry.sdk._logs import LogData
 from opentelemetry.sdk._logs.export import LogExporter
 from opentelemetry.sdk.resources import Resource
 from opentelemetry.trace import format_span_id, format_trace_id
+from opentelemetry.util.types import AnyValue
 
 DEFAULT_MAX_ENTRY_SIZE = 256000  # 256 KB
 DEFAULT_MAX_REQUEST_SIZE = 10000000  # 10 MB
@@ -117,10 +118,37 @@ def _convert_any_value_to_string(value: Any) -> str:
     return ""
 
 
-def _set_payload_in_log_entry(log_entry: LogEntry, body: Any | None):
+# Be careful not to mutate original body. Make copies of anything that needs to change.
+def _sanitized_body(
+    body: Mapping[str, AnyValue]
+) -> MutableMapping[str, AnyValue]:
+    new_body: MutableMapping[str, AnyValue] = {}
+    for key, value in body.items():
+        if (
+            isinstance(value, Sequence)
+            and len(value) > 0
+            and isinstance(value[0], bytes)
+        ):
+            # Should not be possible for a non-bytes value to be present. AnyValue requires Sequence be of one type, and above
+            # we verified the first value is type bytes.
+            new_body[key] = [
+                base64.b64encode(v).decode()
+                for v in value
+                if isinstance(v, bytes)
+            ]
+        elif isinstance(value, bytes):
+            new_body[key] = base64.b64encode(value).decode()
+        elif isinstance(value, Mapping):
+            new_body[key] = _sanitized_body(value)
+        else:
+            new_body[key] = value
+    return new_body
+
+
+def _set_payload_in_log_entry(log_entry: LogEntry, body: AnyValue):
     struct = Struct()
     if isinstance(body, Mapping):
-        struct.update(body)
+        struct.update(_sanitized_body(body))
         log_entry.json_payload = struct
     elif isinstance(body, bytes):
         json_str = body.decode("utf-8", errors="replace")