+++ opentelemetry-resourcedetector-gcp/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+Update opentelemetry-api/sdk dependencies to 1.3.
+
 ## Version 1.9.0a0
 
 Released 2025-02-03